Here's the beginning of the pyfetch script. It barely works, but here's the gist.
Currently, it has scanner device and radioreference.com webservice
authentican information hard coded into it. I plan on moving those to
one or more configuration files.

It uses a wsdl file that I created from the wsdl description at
radioreference.com, but then I had to edit it a bit to make it work.
I'll add that eventually, along with some docs on how I created it.

I should probably eventually break the modules in this script out into
their own libraries so that I can use them from a few tools in this
package. I envision one tool to talk to radioreference.com, one to
talk to the scanner. They should be tied together by a common library
that reads and writes a file that describes what you want in your
scanner. One reason for this is because the information directly from
radioreference.com is not useful for programming your scanner
directly. Everyone will want to edit it a bit. Also, one may want to
use these files to be able to store the state of your scanner so that
you can change it between several configuration options, without
having to hit radioreference.com.