development has stopped
I dropped and broke my scanner, so I'll probably stop working on this
project. I may buy another one or fix this one, but not right now.

I'm also checking in a feature that was half-done in my code. I'm not
really sure if it works or not, but I can't test it now.